+++ botfront/cypress/integration/bot_responses/quickreply_bot_response.spec.js
@@ -61,4 +61,50 @@ describe('Bot responses', function() {
         cy.dataCy('bot-response-input').contains('updated text').should('exist');
         cy.dataCy('button_B').should('exist');
     });
+    it('should create quick reply variations', function() {
+        cy.visit('/project/bf/dialogue/templates');
+        cy.dataCy('create-response').click();
+        cy.dataCy('add-quickreply-response').click();
+        cy.dataCy('response-name-input').click().find('input').type('test_A');
+
+        cy.dataCy('bot-response-input').click().find('textarea').type('response text deleted');
+        cy.dataCy('button_title').click({ force: true });
+        cy.dataCy('button_title').click({ force: true });
+        cy.dataCy('enter-button-title').find('input').type('button deleted');
+        cy.dataCy('intent-label').should('exist').click();
+        cy.dataCy('intent-dropdown').find('input').type('payload{enter}');
+        cy.dataCy('save-button').click();
+
+        cy.dataCy('icon-trash').click();
+
+        cy.dataCy('bot-response-input').contains('response text deleted').should('not.exist');
+        cy.dataCy('button_deleted').should('not.exist');
+
+        cy.dataCy('bot-response-input').click().find('textarea').type('response text A');
+        cy.dataCy('button_title').click({ force: true });
+        cy.dataCy('button_title').click({ force: true });
+        cy.dataCy('enter-button-title').find('input').type('button A');
+        cy.dataCy('intent-label').should('exist').click();
+        cy.dataCy('intent-dropdown').find('input').type('payload{enter}');
+        cy.dataCy('save-button').click();
+
+        cy.dataCy('add-variation').click();
+
+        cy.dataCy('bot-response-input').last().click().find('textarea')
+            .type('response text B');
+        cy.dataCy('button_title').last().click({ force: true });
+        cy.dataCy('button_title').last().click({ force: true });
+        cy.dataCy('enter-button-title').find('input').type('button B');
+        cy.dataCy('intent-label').should('exist').click();
+        cy.dataCy('intent-dropdown').find('input').type('payload{enter}');
+        cy.dataCy('save-button').click();
+        cy.get('.dimmer').click({ position: 'topLeft' }); // close the response editor
+        cy.get('.dimmer').should('not.exist');
+
+        cy.dataCy('edit-response-0').click();
+        cy.dataCy('bot-response-input').contains('response text A').should('exist');
+        cy.dataCy('button_B').should('exist');
+        cy.dataCy('bot-response-input').contains('response text B').should('exist');
+        cy.dataCy('button_A').should('exist');
+    });
 });